requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiogram==3.4.1
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1
web3==6.11.3
flask==2.3.3
xgboost==2.0.1
numpy>=1.26.0
setuptools>=65.0.0
eth-account==0.9.0
eth-utils==2.3.0 